Move fields to table.rs
